moved regular expressions to separate file to organize & cut down size of core.py
