+++ davos/core/core.pyi
@@ -164,15 +164,3 @@ def smuggle(
         installer_kwargs: Optional[PipInstallerKwargs] = ...
 ) -> None: ...
 def smuggle_parser(line: str) -> str: ...
-
-_name_re: Literal[r'[a-zA-Z]\w*']
-
-# noinspection PyPep8Naming
-class _smuggle_subexprs(TypedDict):
-    as_re: Literal[r' +as +[a-zA-Z]\w*']
-    comment_re: Literal[r'(?m:\#+.*$)']
-    name_re: Literal[r'[a-zA-Z]\w*']
-    onion_re: Literal[r'\# *(?:pip|conda) *: *[^#\n ].+?(?= +\#| *\n| *$)']
-    qualname_re: Literal[r'[a-zA-Z]\w*(?: *\. *[a-zA-Z]\w*)*']
-
-smuggle_statement_regex: Pattern[str]
